Only watch for changes in app/packs
Having source_path set to app meant that all files under app/** would be
watched by Webpack devserver, triggering a recompile of assets even if
we changed a model or a view.

This does mean we had to move the actual pack entries to somewhere else,
but that's a small price to pay for not having to wait ~15 seconds on
each page refresh.