Don't assign reviewers
My list of pull requests awaiting code review is getting too long. I'd
usually prefer to focus on reviews from people as they might be blocked
in development.

Bot-generated pull requests are usually not as time-sensitive, so we
should be able to review them at our own pace - usually as part of a
maintenance agreement day.